refactor: unify firefly parsing
