+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/aquasecurity/libbpfgo v0.8.0-libbpf-1.5
 	github.com/containerd/cgroups v1.1.0
 	github.com/cpuguy83/go-md2man/v2 v2.0.4
+	github.com/fatih/structs v1.1.0
 	github.com/labstack/echo/v4 v4.12.0
 	github.com/pion/stun/v3 v3.0.0
 	github.com/pion/transport/v3 v3.0.7
