+++ types/firefly_test.go
@@ -1,19 +1,12 @@
-package fireflyp
+package types
 
 import (
-	"fmt"
 	"os"
 	"path"
 	"testing"
 )
 
 func TestParseIncomingFirefly(t *testing.T) {
-	hasSyslogHeader := map[string]bool{
-		"start_firefly.firefly": false,
-		"end_firefly.firefly":   false,
-		"sample.firefly":        true,
-	}
-
 	testFiles, err := os.ReadDir("testdata")
 	if err != nil {
 		t.Errorf("couldn't list the testdata directory: %v", err)
@@ -27,11 +20,11 @@ func TestParseIncomingFirefly(t *testing.T) {
 			t.Errorf("couldn't read %q: %v", t.Name(), err)
 		}
 
-		firefly, err := parseFirefly(data, hasSyslogHeader[tf.Name()])
-		if err != nil {
+		sFirefly := SlimFirefly{}
+		if err := sFirefly.Parse(data); err != nil {
 			t.Errorf("couldn't parse %q: %v", tfPath, err)
 		}
 
-		fmt.Printf("parsed firefly %q: %+v\n", tfPath, firefly)
+		// fmt.Printf("parsed firefly %q: %+v\n", tfPath, sFirefly)
 	}
 }
